Move _setCookieAcceptPolicy to WKHTTPCookieStore https://bugs.webkit.org/show_bug.cgi?id=216375
Reviewed by Geoffrey Garen.

Source/WebCore:

* platform/network/cocoa/HTTPCookieAcceptPolicyCocoa.h:

Source/WebKit:

It is only used by Safari, which I am changing along with this in rdar://problem/68659937
I need it to not be on the process pool for https://bugs.webkit.org/show_bug.cgi?id=216041
and because the cookie store's job is to manage cookie storage.  The fact that it was on the process pool
is just a historical artifact of us trying to make it global.

WKContextUseTestingNetworkSession has had its important parts replaced by WKWebsiteDataStoreConfigurationSetTestingSessionEnabled
but it was still being used to determinte whether the UI process should persist the cookie accept policy.  We don't need this
any more because we send a message to the network process and let it manage the cookie accept policy.

* UIProcess/API/C/WKContext.cpp:
(WKContextUseTestingNetworkSession): Deleted.
* UIProcess/API/C/WKContextPrivate.h:
* UIProcess/API/Cocoa/WKHTTPCookieStore.mm:
(-[WKHTTPCookieStore _setCookieAcceptPolicy:completionHandler:]):
* UIProcess/API/Cocoa/WKHTTPCookieStorePrivate.h:
* UIProcess/API/Cocoa/WKProcessPool.mm:
(-[WKProcessPool _setCookieAcceptPolicy:]):
* UIProcess/API/Cocoa/WKProcessPoolPrivate.h:
* UIProcess/WebCookieManagerProxy.cpp:
(WebKit::WebCookieManagerProxy::setHTTPCookieAcceptPolicy):
* UIProcess/WebProcessPool.cpp:
(WebKit::WebProcessPool::useTestingNetworkSession): Deleted.
* UIProcess/WebProcessPool.h:

Tools:

* TestWebKitAPI/Tests/WebKitCocoa/CookieAcceptPolicy.mm:
(TEST):
* TestWebKitAPI/Tests/WebKitCocoa/ProcessSwapOnNavigation.mm:
* WebKitTestRunner/TestController.cpp:
(WTR::TestController::generatePageConfiguration):



Identifier: 229249@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266951 268f45cc-cd09-0410-ab3c-d52691b4dbfc