REGRESSION (r266817): ASSERTION FAILED: injectedBundle.isTestRunning() on imported/w3c/web-platform-tests/service-workers/service-worker/svg-target-reftest.https.html https://bugs.webkit.org/show_bug.cgi?id=216440
[GTK] imported/w3c/web-platform-tests/service-workers/service-worker/svg-target-reftest.https.html is flaky
https://bugs.webkit.org/show_bug.cgi?id=210375

Reviewed by Sam Weinig.

LayoutTests/imported/w3c:

* web-platform-tests/service-workers/service-worker/resources/svg-target-reftest-frame.html: Added.
* web-platform-tests/service-workers/service-worker/resources/svg-target-reftest-001-frame.html: Added.
Somehow these files were missed in the web-platform-tests import, and they are both used by the
svg-target-reftest.https.html test.

* web-platform-tests/service-workers/service-worker/svg-target-reftest.https-expected.html:
Updated the path in this expected result file. The test importer script should have done this;
not sure why it did not.

Tools:

Did some hardening of the reftest-wait mechanism in WebKitTestRunner.
We can later do the same in DumpRenderTree if the same problem comes up there,
but at this time I don't have a test that reproduces the issue there so not
changing it for now. Don't want to add speculative code without testing it.

* WebKitTestRunner/InjectedBundle/InjectedBundlePage.cpp:
(WTR::hasRefTestWaitAttribute): Changed to take a WKBundlePageRef instead of
a InjectedBundlePage&.
(WTR::dumpAfterWaitAttributeIsRemoved): Changed to take a WKBundlePageRef.
Added code to handle cases where an explicit dump occurs while we are waiting
for the attribute to be removed. Identify the page with the WKBundlePageRef,
which can be retained/released.
(WTR::InjectedBundlePage::frameDidChangeLocation): Check page for null
instead of indirectly checking by asking if the pageCount is 0. Also pass
a WKBundlePageRef in to dumpAfterWaitAttributeIsRemoved.

LayoutTests:

* TestExpectations: Skip svg-target-reftest.https.html because it fails in
legacy WebKit and the symptom is a hang waiting for the reftest-wait attribute
to be removed. The failure is not new.

* platform/gtk/TestExpectations: Removed expectation that this same test will
be flaky. The addition of reftest-wait support and the subresource along with
the bug fix to WebKitTestRunner should leave this passing and non-flaky on GTK.
Feel free to add this back if that proves wrong.

* platform/wk2/TestExpectations: Expect a pass for this test. It's working fine
on modern WebKit.


Identifier: 229274@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266984 268f45cc-cd09-0410-ab3c-d52691b4dbfc