Fix root cause of problem with text replacements at the beginnings of sentences so we can remove the workaround https://bugs.webkit.org/show_bug.cgi?id=216450
Reviewed by Tim Horton.

* dom/SimpleRange.cpp:
(WebCore::compareByDocumentOrder): Added so we can use std::min/max.
(WebCore::unionRange): Refactored to use std::min/max.
(WebCore::intersection): Added.
* dom/SimpleRange.h: Added intersection.

* editing/Editor.cpp:
(WebCore::Editor::markMisspellingsAfterTypingToWord): Intersect with the sentence range,
since we don't have a guarantee that moving to the start of a word and end of a word
stays within the sentence since the word and sentence algorithms are not wholly consistent.
In such cases we would like to restrict our checking to the sentence. Another way to fix
this would be to change the behavior of the word and sentence boundary functions.
(WebCore::Editor::markAndReplaceFor): Remove the boolean "true" flags that we used to
trigger our workaround for the above.

* editing/TextCheckingHelper.cpp:
(WebCore::TextCheckingParagraph::automaticReplacementStart const): Removed the boolean
argument and the workaround code that runs when it's true.
(WebCore::TextCheckingParagraph::automaticReplacementLength const): Ditto.
* editing/TextCheckingHelper.h: Removed the boolean arguments.

Identifier: 229277@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@266987 268f45cc-cd09-0410-ab3c-d52691b4dbfc