DFG should ensure there are PhantomLocals for the taken block of op_jneq_ptr https://bugs.webkit.org/show_bug.cgi?id=216669
Reviewed by Saam Barati.

JSTests:

* stress/jneq-ptr-opcode-variable-only-live-on-taken-branch.js: Added.
(bar):
(foo):

Source/JavaScriptCore:

Right now, if there is a local that is live on the taken branch but dead on
not-taken branch then nothing will preserve it for OSR exit. This patch simply
adds a PhantomLocal for each live operand for the first bytecode of the taken block.

* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::parseBlock):


Identifier: 229514@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267255 268f45cc-cd09-0410-ab3c-d52691b4dbfc